requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
arxiv==2.1.0
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.2
black==23.12.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==41.0.7
cssselect==1.2.0
curl-cffi==0.5.10
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.9.0
duckduckgo_search==4.1.1
feedparser==6.0.10
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
hyperlink==21.0.0
idna==3.6
importlib-metadata==6.11.0
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
llama-index
lxml==5.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.3
openai==1.6.1
outcome==1.3.0.post0
pandas==2.1.4
parsel==1.8.1
pathspec==0.12.1
pillow==10.2.0
Protego==0.3.0
protobuf==4.25.1
pyarrow==14.0.2
pyarxiv==1.0.3.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
PyDispatcher==2.0.7
pyOpenSSL==23.3.0
pypdf==3.17.4
PyPDF2==3.0.1
pyprojroot==0.3.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
queuelib==1.6.2
rake-nltk==1.0.6
readability-lxml==0.8.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-file==1.5.1
rich==13.7.0
rpds-py==0.16.2
Scrapy==2.11.0
selenium==4.16.0
service-identity==23.1.0
setuptools==69.0.3
sgmllib3k==1.0.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
streamlit==1.29.0
streamlit-chat==0.1.1
tenacity==8.2.3
tiktoken==0.5.2
tldextract==5.1.1
toml==0.10.2
toolz==0.12.0
tqdm==4.66.1
trio==0.23.2
trio-websocket==0.11.1
Twisted==22.10.0
typing-inspect==0.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
w3lib==2.1.2
watchdog==3.0.0
webdriver-manager==4.0.1
wheel==0.42.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
youtube-transcript-api==0.6.2
zope.interface==6.1
llama-parse 